requirements.txt
streamlit
tensorflow-cpu
opencv-python-headless
pillow
pandas
plotly